refactor logic of structure

+++ src/main/java/oqu/today/initital/repository/DBFileRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface DBFileRepository extends JpaRepository<DBFile, String> {
-    public void deleteByUserId(int userId);
+    public void deleteByUserId(long userId);
 }
